pyproject.toml
[project]
name = "nemoslides"
version = "0.1.0"
description = "NemoSlides: fine-tune Nemotron-3-Nano-30B-A3B to generate Slidev decks. NVIDIA Nemotron Hackathon 2026, Track B."
requires-python = ">=3.11"
dependencies = [
    "openai>=1.54.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "tqdm>=4.66.0",
    "datasets>=4.8.4",
    "huggingface-hub>=1.11.0",
    "data-designer>=0.2.1",
    "fastapi>=0.136.0",
    "uvicorn>=0.45.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.26",
]

[project.urls]
Repository = "https://github.com/trillion-labs/nemoslides"
Documentation = "https://trillion-labs.github.io/nemoslides/"
Dataset = "https://huggingface.co/datasets/trillionlabs/slides-sft-v0"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.6.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "pymdown-extensions>=10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nemoslides"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "src"]
